Add macro to define linked associations
