+++ lib/pco_api/people/address.ex
@@ -1,4 +1,3 @@
 defmodule PcoApi.People.Address do
-  use PcoApi.ActionsTwo
-  defstruct [:attributes, :id, :links, :type]
+  use PcoApi.Actions
 end
