+++ lib/pco_api/people/household.ex
@@ -1,4 +1,3 @@
 defmodule PcoApi.People.Household do
-  use PcoApi.ActionsTwo
-  defstruct [:attributes, :id, :links, :type]
+  use PcoApi.Actions
 end
